requirements.txt
# Django и основные компоненты
Django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.3.1

# База данных
psycopg2-binary==2.9.9  # для PostgreSQL в продакшене

# Работа с изображениями
Pillow==10.4.0

# QR-коды
qrcode[pil]==7.4.2

# Валидация и сериализация
django-phonenumber-field==7.3.0
phonenumbers==8.13.40

# Аутентификация и безопасность
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Утилиты для разработки
python-decouple==3.8  # для переменных окружения
django-extensions==3.2.3  # полезные команды Django

# Кеширование
django-redis==5.4.0

# Файлы и медиа
django-storages==1.14.2  # для AWS S3
boto3==1.34.144  # AWS SDK

# Тестирование
factory-boy==3.3.0
pytest-django==4.8.0
pytest-cov==4.1.0

# Продакшен
gunicorn==21.2.0
whitenoise==6.6.0

# Мониторинг и логи
sentry-sdk==2.8.0

# API документация
drf-spectacular==0.27.2 